pyproject.toml
[tool.poetry]
name = "transaction-api-test-task"
version = "0.1.0"
description = "API service of the management of Transaction API written on DRF"
authors = ["Viktor Beniukh <superbigbenjamin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.6"
python-dotenv = "^1.0.1"
django-rest-framework = "^0.1.0"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
